djangogirls/requirements.txt
Django==1.9.4
djangorestframework==3.4.7
